TanitaScale/src/ca/charland/tanita/BoneMassActivity.java
package ca.charland.tanita;

import ca.charland.tanita.db.TanitaDataTable;

/**
 * The Class BoneMassActivity.
 * 
 * @author mcharland
 */
public class BoneMassActivity extends AbstractTextViewActivity {

	/** {@inheritDoc} */
	@Override
	int getLayoutResID() {
		return R.layout.bone_mass;
	}
	
	/** {@inheritDoc} */
	@Override
	public Class<?> getNextClass() {
		return BodyFatTotalActivity.class;
	}

	/** {@inheritDoc} */
	@Override
	protected TanitaDataTable.Column getColumnName() {
		return TanitaDataTable.Column.BONE_MASS;
	}
}
